% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacceptprob.R
\name{bacceptprob}
\alias{bacceptprob}
\title{Acceptance Probability for Binomial RDT}
\usage{
bacceptprob(n, c, pi)
}
\arguments{
\item{n}{RDT sample size.}

\item{c}{Maximum allowable failures.}

\item{pi}{Failure probability.}
}
\value{
Acceptance probability
}
\description{
Define the acceptance probability function which gets the probability of passing the test (for binomial RDT).
}
\examples{
pi <- pi_MCSim_beta(M = 5000, seed = 10, a = 1, b = 1)
bacceptprob(n = 10, c = 2, pi = pi);
}
