% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcost_RG.R
\name{bcost_RG}
\alias{bcost_RG}
\title{Reliability Growth Cost}
\usage{
bcost_RG(G)
}
\arguments{
\item{G}{A constant value reliabilty growth cost, suggest to be sufficiently larger than RDT cost.}
}
\value{
Reliability growth cost
}
\description{
Define the cost function of reliabiltiy growth (RG) after the decision of the test (for binomial RDT).
}
\examples{
bcost_RG(G = 100000);
}
\seealso{
\code{\link{bcost_RDT}}, \code{\link{bcost_WS}}, \code{\link{bcost_expected}}
}
